bench: add criterion benchmarks
This represents an initial go at adding comprehensive criterion
benchmarks. We hit every public API function, and vary benchmarks based
on both corpus size and match frequency.

The point of this exercise is to establish a baseline on the status quo,
and make sure we don't introduce any major regressions after a
refactoring toward vendor intrinsics.