memchr: rewrite everything
This fills out the complete API of the memchr crate, namely, including
memrchr2 and memrchr3, along with doubled ended iterators for memchr2
and memchr3. We also add `memr?chr[123]_iter` convenience free functions
for constructing iterators, instead of needing to import the iterator
type itself.

In doing so, we provide vectorized implementations for everything. Both
memchr and memrchr should now be competitive with glibc on x86_64, and
therefore, we now drop libc on x86_64. This also has the benefit of
providing consistent performance across all platforms, regardless of
which libc is used. Namely, the performance of the various
implementations of memchr across different implementations of libc can
vary greatly.

We also rewrite the fallback routines to make them faster as well.
For the most part, throughput remains roughly the same, but we reduce
overhead quite a bit by just accepting our fate and using raw pointers
everywhere.

Closes #19, Closes #34