simd: speed up position extraction
By simplifying forward_pos and reverse_pos by ORing masks.
This works because we always want the first matching position,
and ORing the masks won't change the result with respect to
extracting the number of trailing zeroes. (And similarly for the
reverse direction.)

This has the highest impact on cases where memchr will match
a lot, which is an important case to make fast.