build: only enable simd on targets which feature SSE
While odd, it is possible for x86_64 to not have SSE enabled. In that
case, we want to avoid enabling the explicit SIMD options, as it will
result in a compilation failure. We address this by checking the CPU
target features at compile time. If SSE isn't in them, then we don't
compile the SIMD stuff.

We don't need to do this for SSE4.2 or AVX since they are selected at
runtime. For SSE2, the code assumes it exists for x86_64.

Fixes #57, Closes #77